+++ tloen/domain/__init__.py
@@ -7,10 +7,11 @@ from .bases import (
     Container,
 )
 from .chains import Chain, ChainContainer, RackDevice, Transfer
-from .clips import Clip, Envelope, Note, NoteMoment, Scene, Slot, Timeline
+from .clips import Clip, Note, NoteMoment
 from .contexts import Context
 from .controllers import Controller
 from .devices import DeviceIn, DeviceObject, DeviceOut
+from .envelops import Envelope
 from .instruments import BasicSampler, BasicSynth, Instrument
 from .midieffects import Arpeggiator, Chord
 from .parameters import (
@@ -23,6 +24,7 @@ from .parameters import (
     ParameterObject,
 )
 from .sends import DirectIn, DirectOut, Patch, Send, SendObject, Target
+from .slots import Scene, Slot
 from .tracks import (
     CueTrack,
     MasterTrack,
@@ -75,7 +77,6 @@ __all__ = [
     "SendObject",
     "Slot",
     "Target",
-    "Timeline",
     "Track",
     "TrackContainer",
     "TrackObject",
