+++ tloen/domain/envelopes.py
@@ -1,0 +1,6 @@
+class Envelope:
+    """
+    An automation envelope, in a Clip or Timeline.
+    """
+
+    pass
