+++ tloen/domain/tracks.py
@@ -1,7 +1,7 @@
 import abc
 import logging
 from types import MappingProxyType
-from typing import Dict, List, Mapping, Optional, Set, Type, Union
+from typing import Dict, Mapping, Optional, Set, Type, Union
 from uuid import UUID, uuid4
 
 from supriya.enums import AddAction, CalculationRate
@@ -17,10 +17,11 @@ from .bases import (
     Mixer,
     Performable,
 )
-from .clips import Clip, ClipLaunched, Scene, Slot
+from .clips import Clip, ClipContainer
 from .devices import DeviceObject
 from .parameters import BusParameter, Float, ParameterGroup, ParameterObject
 from .sends import Receive, Send, Target
+from .slots import ClipLaunched, Scene, Slot
 from .synthdefs import build_patch_synthdef, build_peak_rms_synthdef
 
 logger = logging.getLogger("tloen.domain")
@@ -556,7 +557,7 @@ class Track(UserTrackObject):
         self._clip_launch_event_id: Optional[int] = None
         self._clip_perform_event_id: Optional[int] = None
         self._pending_slot_index: Optional[int] = None
-        self._clips = Container(label="Clips")
+        self._clips = ClipContainer(label="Clips")
         self._slots = Container(label="Slots")
         self._tracks = TrackContainer("input", AddAction.ADD_AFTER, label="SubTracks")
         self._mutate(slice(1, 1), [self._clips, self._slots, self._tracks])
@@ -781,12 +782,6 @@ class Track(UserTrackObject):
             self._tracks._append(track)
             return track
 
-    async def delete_clips(self, *, from_, to):
-        ...
-
-    async def duplicate_clips(self, *, from_, to) -> List[Clip]:
-        ...
-
     @classmethod
     async def group(cls, tracks, *, name=None):
         async with cls.lock(tracks):
@@ -798,8 +793,10 @@ class Track(UserTrackObject):
             group_track.tracks._mutate(slice(None), tracks)
             return group_track
 
-    async def insert_clip(self, *, from_, to) -> Clip:
-        clip = Clip()
+    async def insert_clip(self, *, from_: float, to: float) -> Clip:
+        if to >= from_:
+            raise ValueError
+        clip = Clip(start_offset=from_, stop_offset=to, stop_marker=to - from_,)
         return clip
 
     async def move(self, container, position):
