+++ tloen/gridui/clipview.py
@@ -6,7 +6,8 @@ from monome import GridApp, GridBuffer
 
 from ..bases import Event
 from ..commands.clips import ToggleClipNote
-from ..domain.clips import Clip, ClipLaunched, ClipModified
+from ..domain.clips import Clip, ClipModified
+from ..domain.slots import ClipLaunched
 from ..domain.transports import TransportStopped, TransportTicked
 from ..pubsub import PubSub
